Standardize TFIDF test to test against the interface
Multiple implementations can test against the same test cases and their accuracy can be measured against each other
